Fix: HSMM state time course should be int.

+++ osl_dynamics/simulation/hsmm.py
@@ -127,7 +127,7 @@ class HSMM:
             )
             current_position += state_lifetime
 
-        return alpha
+        return alpha.astype(int)
 
 
 class HSMM_MVN(Simulation):
